pyproject.toml
[project]
name = "crds"
version = "0.1.0"
description = "Pull Kubernetes CRDs from a url and generate json schemas for them"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "argparse==1.4.0",
    "certifi==2026.2.25",
    "charset-normalizer==3.4.6",
    "idna==3.11",
    "jinja2==3.1.6",
    "markupsafe==3.0.3",
    "pytest>=9.0.2",
    "pyyaml==6.0.3",
    "requests==2.32.5",
    "urllib3==2.6.3",
]

[tool.pytest.ini_options]
testpaths = ["src"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
